[102] && [103] refactored code
