+++ src/main/front/components/Modal/Signin/Signin.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 
 /* Components */
-import ModalSkeleton from '../../Modal';
+import ModalSkeleton from '../../Modal/ModalSkeleton';
 import SigninForm from './components/SigninForm.js';
 
 /* Styles */
@@ -12,21 +12,17 @@ import './Signin.less';
 class Signin extends Component {
 	constructor(props) {
 		super(props);
-
-		this.toggleActive = this.toggleActive.bind(this);
 	}
 
-	toggleActive = () => {}
-
 	render() {
-		return(
+		return (
 			<ModalSkeleton isActive={this.props.isActive}>
 				<div id="signin-container" className="modal-content">
-					<span className="closeModal" onClick={this.toggleActive} />
+					<span className="closeModal" />
 					<div className="inner">
 						<h3> Welcome Back </h3>
 						<SigninForm />
-						
+
 						<div className="link-group">
 							<a href="#"> Sign up </a>
 							<a href="#"> Forgot password </a>
