+++ src/main/front/components/Modal/Signin/Signin.less
@@ -42,7 +42,7 @@
 				}
 			}
 
-			// hides formik error messages in forms
+			// TEMP: hides formik error messages in forms
 			> div {
 				display: none;
 			}
