--- src/main/front/components/Modal/Signin/index.js
@@ -1,2 +1,0 @@
-import Signin from './Signin.js';
-export default Signin;
