+++ src/main/front/components/Modal/Signup/Signup.less
@@ -90,7 +90,7 @@
 
 					transition: left .15s linear;
 
-					// hides formik error messages in forms
+					// TEMP: hides formik error messages in forms
 					> div {
 						display: none;
 					}
