--- src/main/front/components/Modal/Signup/index.js
@@ -1,2 +1,0 @@
-import Signup from './Signup.js';
-export default Signup;
