--- src/main/front/components/Modal/index.js
@@ -1,2 +1,0 @@
-import ModalSkeleton from './ModalSkeleton.js';
-export default ModalSkeleton;
