+++ src/main/front/pages/MainPage.js
@@ -2,30 +2,22 @@ import React, { Component } from 'react';
 
 /* Components */
 import Button from "../components/Button/Button";
-import Signin from "../components/Modal/Signin";
-import Signup from "../components/Modal/Signup";
+import Signin from "../components/Modal/Signin/Signin";
+import Signup from "../components/Modal/Signup/Signup";
 
 
 class MainPage extends Component {
-
   constructor(props) {
     super(props);
+    
     this.state = {
       signinOpen: false,
-      signinOpen: false
+      signupOpen: false
     };
-
-    this.toggleSignin = this.toggleSignin.bind(this);
-    this.toggleSignup = this.toggleSignup.bind(this);
   }
 
-  toggleSignin = () => {
-    this.setState({ signinOpen: this.state.signinOpen ? false : true });
-  }
-
-  toggleSignup = () => {
-    this.setState({ signupOpen: this.state.signupOpen ? false : true });
-  }
+  toggleSignin = () => this.setState({ signinOpen: this.state.signinOpen ? false : true });
+  toggleSignup = () => this.setState({ signupOpen: this.state.signupOpen ? false : true });
 
   render() {
     return (
